Criando classe base para previsões
